+++ packages/sdk/src/definition/soap-introspection.ts
@@ -1,8 +1,12 @@
 import { GraphQLApi } from './index';
 import { introspectWithCache } from './introspection-cache';
 import { IHeadersBuilder } from './headers-builder';
-import { soapToGraphQLApi } from '../introspection/soap';
-import { readFile } from './openapi-introspection';
+import { readFile, validateIntrospectionId } from './openapi-introspection';
+import { SOAPLoader } from '@omnigraph/soap';
+import { fetch } from '@whatwg-node/fetch';
+import { printSchemaWithDirectives } from '@graphql-tools/utils';
+import { introspectGraphql } from './graphql-introspection';
+import { WgEnv } from '../configure/options';
 
 export interface SoapIntrospectionFile {
 	kind: 'file';
@@ -27,5 +31,33 @@ export const introspectSoap = async (introspection: SoapIntrospection) => {
 		async (introspection: SoapIntrospection): Promise<GraphQLApi> => {
 			return await soapToGraphQLApi(spec, introspection);
 		}
+	);
+};
+
+const soapToGraphQLApi = async (wsdl: string, introspection: SoapIntrospection): Promise<GraphQLApi> => {
+	const apiID: string = await validateIntrospectionId(introspection.id);
+
+	const soapLoader = new SOAPLoader({
+		fetch,
+	});
+
+	await soapLoader.loadWSDL(wsdl);
+	const graphQLSchema = soapLoader.buildSchema();
+
+	const schema = printSchemaWithDirectives(graphQLSchema);
+
+	return introspectGraphql(
+		{
+			url: `${WgEnv.ServerUrl}-soap`,
+			// baseUrl: introspection.baseURL || '',
+			path: `/soap/${apiID}`,
+			apiNamespace: introspection.apiNamespace,
+			internal: true,
+			loadSchemaFromString: () => schema,
+			headers: introspection.headers,
+			customIntScalars: ['BigInt'],
+		},
+		{},
+		true
 	);
 };
